Add variable skip_subgid_change with default value true
Adding subgids and subguids is not needed anymore since useradd
adds them automatically.

The addition implemented here only works for the first user (id 1000 with
subids starting from 100000).
For other users it leads to overlapping subuid/subgid ranges
which might lead to very confusing issues when running containers.

The addition is left here in case someone is using it or wants
to override default ids with some specific values.

Fixes #34