Added folder creation for volumes (containers, not pods)
First it checks if the folder already exists and if it does, it won't
adjust any permissions. This helps if podman can't manage the
permissions correctly.

It allows for changing the owner and group in case it is needed to set a
specific UID and GID.

It also allows to change the mode.

I added explanations for :U as well, which tells podman to change the
permissions to the container user recuresively. This works if the
service inside the container doesn't run with a different user than the
container.